change np32 to np64

+++ utils/utils.py
@@ -13,7 +13,7 @@ def load_obj(folder, name):
     with open(path, 'rb') as f:
         return pickle.load(f)
 
-zelda_spaces = (Box(low=0, high=1, shape=(13, 9, 13), dtype=np.float32), Discrete(6))
+zelda_spaces = (Box(low=0, high=1, shape=(13, 9, 13), dtype=np.float64), Discrete(6))
 
 def add_noise(grid):
     """add noise to the grid. Let that noise be flip the bit from 0 to 1 or vice-versa
